% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_bcc.R
\name{bcc_cols}
\alias{bcc_cols}
\title{Extract BCC colors as hex codes}
\usage{
bcc_cols(...)
}
\arguments{
\item{...}{character Names of bcc_colours. If empty the full list of colours is returned.}
}
\value{
A vector of hex codes for the specified colour names.
}
\description{
Extract BCC colors as hex codes
}
\examples{
\dontrun{
bcc_cols()
bcc_cols("pink", "white")
bcc_cols("yellow", "green", "blue")
}

}
