% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_bcc.R
\name{bcc_pal}
\alias{bcc_pal}
\title{Interpolate a BCC colour palette}
\usage{
bcc_pal(palette = "purple", reverse = FALSE, ...)
}
\arguments{
\item{palette}{string Name of palette in bcc_palettes. Defaults to "purple".}

\item{reverse}{boolean Should the palette should be reversed? Defaults to FALSE.}

\item{...}{Additional arguments to pass to colorRampPalette().}
}
\value{
A function to interpolate the colours present in the palette.
}
\description{
Interpolate a BCC colour palette
}
\examples{
\dontrun{
bcc_pal("pink")(10)
bcc_pal("multi")(25)
}

}
