% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_bcc.R
\name{scale_colour_bcc}
\alias{scale_colour_bcc}
\title{Color scale constructor for BCC colours}
\usage{
scale_colour_bcc(palette = "purple", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{string Name of palette in bcc_palettes}

\item{discrete}{boolean Is the colour aesthetic discrete? Defaults to TRUE.}

\item{reverse}{boolean Should the palette should be reversed? Defaults to FALSE.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
A colour scale function for ggplot2.
}
\description{
Color scale constructor for BCC colours
}
\examples{
\dontrun{
library(ggplot2)
iris <- data(iris)

p <- ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
geom_point(size = 4)
p + scale_colour_bcc(discrete = FALSE, palette = "orange", reverse = TRUE, guide = "none")
}

}
