% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_bcc.R
\name{scale_fill_bcc}
\alias{scale_fill_bcc}
\title{Fill scale constructor for BCC colours}
\usage{
scale_fill_bcc(palette = "purple", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{string Name of palette in bcc_palettes}

\item{discrete}{boolean Is the fill aesthetic discrete? Defaults to TRUE.}

\item{reverse}{boolean Should the palette should be reversed? Defaults to FALSE.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
A fill scale function for ggplot2.
}
\description{
Fill scale constructor for BCC colours
}
\examples{
\dontrun{
library(ggplot2)
iris <- data(iris)

p <- ggplot(iris, aes(Species, Sepal.Width, fill = Species)) +
geom_col()
p + scale_fill_bcc("multi", guide = "none")
}

}
