% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bcc.R
\name{theme_bcc}
\alias{theme_bcc}
\title{Creates the BCC Theme for ggplot}
\usage{
theme_bcc(
  base_family = "Verdana",
  base_size = 16,
  gridline_x = T,
  gridline_y = T,
  legend_position = "right"
)
}
\arguments{
\item{base_family}{The font to use in this theme. Defaults to "Verdana"}

\item{base_size}{Base font size, in pts. Defaults to 16}

\item{gridline_x}{Should horizontal gridlines be applied? Defaults to TRUE}

\item{gridline_y}{Should vertical gridlines be applied? Defaults to TRUE}

\item{legend_position}{Position of legend in plot. One of "none", "top", "right", "bottom", "left" or "inside" - defaults to "right"}
}
\description{
Creates a base theme for ggplot objects.
}
